+++ lib/common/output.c
@@ -28,46 +28,48 @@ double yDir (double y)
     return YDIR(y);
 }
 
-static int (*putstr) (void *chan, const char *str);
-
-static void agputs (const char* s, FILE* fp)
-{
+static void agputs(int (*putstr)(void *chan, const char *str), const char* s,
+                   FILE* fp) {
     putstr(fp, s);
 }
 
-static void agputc(char c, FILE *fp) {
+static void agputc(int (*putstr)(void *chan, const char *str), char c,
+                   FILE *fp) {
     static char buf[2] = {'\0','\0'};
     buf[0] = c;
     putstr(fp, buf);
 }
 
-static void printstring(FILE * f, char *prefix, char *s)
-{
-    if (prefix) agputs(prefix, f);
-    agputs(s, f);
+static void printstring(int (*putstr)(void *chan, const char *str), FILE *f,
+                        char *prefix, char *s) {
+  if (prefix) agputs(putstr, prefix, f);
+  agputs(putstr, s, f);
 }
 
-static void printint(FILE *f, char *prefix, size_t i) {
-    char buf[BUFSIZ];
+static void printint(int (*putstr)(void *chan, const char *str), FILE *f,
+                     char *prefix, size_t i) {
+  agxbuf buf = {0};
     
-    if (prefix) agputs(prefix, f);
-    snprintf(buf, sizeof(buf), "%" PRISIZE_T, i);
-    agputs(buf, f);
+  if (prefix) agputs(putstr, prefix, f);
+  agxbprint(&buf, "%" PRISIZE_T, i);
+  agputs(putstr, agxbuse(&buf), f);
+  agxbfree(&buf);
 }
 
-static void printdouble(FILE * f, char *prefix, double v)
-{
-    char buf[BUFSIZ];
+static void printdouble(int (*putstr)(void *chan, const char *str), FILE *f,
+                        char *prefix, double v) {
+  agxbuf buf = {0};
     
-    if (prefix) agputs(prefix, f);
-    snprintf(buf, sizeof(buf), "%.5g", v);
-    agputs(buf, f);
+  if (prefix) agputs(putstr, prefix, f);
+  agxbprint(&buf, "%.5g", v);
+  agputs(putstr, agxbuse(&buf), f);
+  agxbfree(&buf);
 }
 
-static void printpoint(FILE * f, pointf p)
-{
-    printdouble(f, " ", PS2INCH(p.x));
-    printdouble(f, " ", PS2INCH(YDIR(p.y)));
+static void printpoint(int (*putstr)(void *chan, const char *str), FILE *f,
+                       pointf p) {
+  printdouble(putstr, f, " ", PS2INCH(p.x));
+  printdouble(putstr, f, " ", PS2INCH(YDIR(p.y)));
 }
 
 /* setYInvert:
@@ -95,15 +97,16 @@ static char* canon (graph_t *g, char* s)
     return cs;
 }
 
-static void writenodeandport(FILE *f, node_t *node, char *portname) {
+static void writenodeandport(int (*putstr)(void *chan, const char *str),
+                             FILE *f, node_t *node, char *portname) {
     char *name;
     if (IS_CLUST_NODE(node))
 	name = canon (agraphof(node), strchr(agnameof(node), ':') + 1);
     else
 	name = agcanonStr (agnameof(node));
-    printstring(f, " ", name); /* slimey i know */
+    printstring(putstr, f, " ", name); /* slimey i know */
     if (portname && *portname)
-	printstring(f, ":", agcanonStr(portname));
+	printstring(putstr, f, ":", agcanonStr(portname));
 }
 
 void write_plain(GVJ_t *job, graph_t *g, FILE *f, bool extend) {
@@ -115,33 +118,33 @@ void write_plain(GVJ_t *job, graph_t *g, FILE *f, bool extend) {
     char *lbl;
     char* fillcolor;
 
-    putstr = g->clos->disc.io->putstr;
+    int (*putstr)(void *chan, const char *str) = g->clos->disc.io->putstr;
     setYInvert(g);
     pt = GD_bb(g).UR;
-    printdouble(f, "graph ", job->zoom);
-    printdouble(f, " ", PS2INCH(pt.x));
-    printdouble(f, " ", PS2INCH(pt.y));
-    agputc('\n', f);
+    printdouble(putstr, f, "graph ", job->zoom);
+    printdouble(putstr, f, " ", PS2INCH(pt.x));
+    printdouble(putstr, f, " ", PS2INCH(pt.y));
+    agputc(putstr, '\n', f);
     for (n = agfstnode(g); n; n = agnxtnode(g, n)) {
 	if (IS_CLUST_NODE(n))
 	    continue;
-	printstring(f, "node ", agcanonStr(agnameof(n)));
-	printpoint(f, ND_coord(n));
+	printstring(putstr, f, "node ", agcanonStr(agnameof(n)));
+	printpoint(putstr, f, ND_coord(n));
 	if (ND_label(n)->html)   /* if html, get original text */
 	    lbl = agcanonStr (agxget(n, N_label));
 	else
 	    lbl = canon(agraphof(n),ND_label(n)->text);
-        printdouble(f, " ", ND_width(n));
-        printdouble(f, " ", ND_height(n));
-        printstring(f, " ", lbl);
-	printstring(f, " ", late_nnstring(n, N_style, "solid"));
-	printstring(f, " ", ND_shape(n)->name);
-	printstring(f, " ", late_nnstring(n, N_color, DEFAULT_COLOR));
+        printdouble(putstr, f, " ", ND_width(n));
+        printdouble(putstr, f, " ", ND_height(n));
+        printstring(putstr, f, " ", lbl);
+	printstring(putstr, f, " ", late_nnstring(n, N_style, "solid"));
+	printstring(putstr, f, " ", ND_shape(n)->name);
+	printstring(putstr, f, " ", late_nnstring(n, N_color, DEFAULT_COLOR));
 	fillcolor = late_nnstring(n, N_fillcolor, "");
         if (fillcolor[0] == '\0')
 	    fillcolor = late_nnstring(n, N_color, DEFAULT_FILL);
-	printstring(f, " ", fillcolor);
-	agputc('\n', f);
+	printstring(putstr, f, " ", fillcolor);
+	agputc(putstr, '\n', f);
     }
     for (n = agfstnode(g); n; n = agnxtnode(g, n)) {
 	for (e = agfstout(g, n); e; e = agnxtout(g, e)) {
@@ -160,26 +163,26 @@ void write_plain(GVJ_t *job, graph_t *g, FILE *f, bool extend) {
 		    bz = ED_spl(e)->list[i];
 		    splinePoints += bz.size;
 		}
-		printstring(f, NULL, "edge");
-		writenodeandport(f, agtail(e), tport);
-		writenodeandport(f, aghead(e), hport);
-		printint(f, " ", splinePoints);
+		printstring(putstr, f, NULL, "edge");
+		writenodeandport(putstr, f, agtail(e), tport);
+		writenodeandport(putstr, f, aghead(e), hport);
+		printint(putstr, f, " ", splinePoints);
 		for (size_t i = 0; i < ED_spl(e)->size; i++) {
 		    bz = ED_spl(e)->list[i];
 		    for (size_t j = 0; j < bz.size; j++)
-			printpoint(f, bz.list[j]);
+			printpoint(putstr, f, bz.list[j]);
 		}
 	    }
 	    if (ED_label(e)) {
-		printstring(f, " ", canon(agraphof(agtail(e)),ED_label(e)->text));
-		printpoint(f, ED_label(e)->pos);
+		printstring(putstr, f, " ", canon(agraphof(agtail(e)),ED_label(e)->text));
+		printpoint(putstr, f, ED_label(e)->pos);
 	    }
-	    printstring(f, " ", late_nnstring(e, E_style, "solid"));
-	    printstring(f, " ", late_nnstring(e, E_color, DEFAULT_COLOR));
-	    agputc('\n', f);
+	    printstring(putstr, f, " ", late_nnstring(e, E_style, "solid"));
+	    printstring(putstr, f, " ", late_nnstring(e, E_color, DEFAULT_COLOR));
+	    agputc(putstr, '\n', f);
 	}
     }
-    agputs("stop\n", f);
+    agputs(putstr, "stop\n", f);
 }
 
 static void set_record_rects(node_t * n, field_t * f, agxbuf * xb)
@@ -200,31 +203,28 @@ static void set_record_rects(node_t * n, field_t * f, agxbuf * xb)
 static void rec_attach_bb(graph_t * g, Agsym_t* bbsym, Agsym_t* lpsym, Agsym_t* lwsym, Agsym_t* lhsym)
 {
     int c;
-    char buf[BUFSIZ];
+    agxbuf buf = {0};
     pointf pt;
 
-    snprintf(buf, sizeof(buf), "%.5g,%.5g,%.5g,%.5g", GD_bb(g).LL.x,
-             YDIR(GD_bb(g).LL.y), GD_bb(g).UR.x, YDIR(GD_bb(g).UR.y));
-    agxset(g, bbsym, buf);
+    agxbprint(&buf, "%.5g,%.5g,%.5g,%.5g", GD_bb(g).LL.x, YDIR(GD_bb(g).LL.y),
+              GD_bb(g).UR.x, YDIR(GD_bb(g).UR.y));
+    agxset(g, bbsym, agxbuse(&buf));
     if (GD_label(g) && GD_label(g)->text[0]) {
 	pt = GD_label(g)->pos;
-	snprintf(buf, sizeof(buf), "%.5g,%.5g", pt.x, YDIR(pt.y));
-	agxset(g, lpsym, buf);
+	agxbprint(&buf, "%.5g,%.5g", pt.x, YDIR(pt.y));
+	agxset(g, lpsym, agxbuse(&buf));
 	pt = GD_label(g)->dimen;
-	snprintf(buf, sizeof(buf), "%.2f", PS2INCH(pt.x));
-	agxset (g, lwsym, buf);
-	snprintf(buf, sizeof(buf), "%.2f", PS2INCH(pt.y));
-	agxset (g, lhsym, buf);
+	agxbprint(&buf, "%.2f", PS2INCH(pt.x));
+	agxset(g, lwsym, agxbuse(&buf));
+	agxbprint(&buf, "%.2f", PS2INCH(pt.y));
+	agxset(g, lhsym, agxbuse(&buf));
     }
     for (c = 1; c <= GD_n_cluster(g); c++)
 	rec_attach_bb(GD_clust(g)[c], bbsym, lpsym, lwsym, lhsym);
+    agxbfree(&buf);
 }
 
-void attach_attrs_and_arrows(graph_t* g, int* sp, int* ep)
-{
-    int e_arrows;		/* graph has edges with end arrows */
-    int s_arrows;		/* graph has edges with start arrows */
-    char buf[BUFSIZ];		/* Used only for small strings */
+void attach_attrs_and_arrows(graph_t *g, bool *sp, bool *ep) {
     node_t *n;
     edge_t *e;
     pointf ptf;
@@ -235,7 +235,8 @@ void attach_attrs_and_arrows(graph_t* g, int* sp, int* ep)
     Agsym_t* lhsym = NULL;
 
     gv_fixLocale (1);
-    e_arrows = s_arrows = 0;
+    bool e_arrows = false; // graph has edges with end arrows
+    bool s_arrows = false; // graph has edges with start arrows
     setYInvert(g);
     agxbuf xb = {0};
     safe_dcl(g, AGNODE, "pos", "");
@@ -269,18 +270,17 @@ void attach_attrs_and_arrows(graph_t* g, int* sp, int* ep)
 	    }
 	    agset(n, "pos", agxbuse(&xb));
 	} else {
-	    snprintf(buf, sizeof(buf), "%.5g,%.5g", ND_coord(n).x,
-	             YDIR(ND_coord(n).y));
-	    agset(n, "pos", buf);
+	    agxbprint(&xb, "%.5g,%.5g", ND_coord(n).x, YDIR(ND_coord(n).y));
+	    agset(n, "pos", agxbuse(&xb));
 	}
-	snprintf(buf, sizeof(buf), "%.5g", PS2INCH(ND_ht(n)));
-	agxset(n, N_height, buf);
-	snprintf(buf, sizeof(buf), "%.5g", PS2INCH(ND_lw(n) + ND_rw(n)));
-	agxset(n, N_width, buf);
+	agxbprint(&xb, "%.5g", PS2INCH(ND_ht(n)));
+	agxset(n, N_height, agxbuse(&xb));
+	agxbprint(&xb, "%.5g", PS2INCH(ND_lw(n) + ND_rw(n)));
+	agxset(n, N_width, agxbuse(&xb));
 	if (ND_xlabel(n) && ND_xlabel(n)->set) {
 	    ptf = ND_xlabel(n)->pos;
-	    snprintf(buf, sizeof(buf), "%.5g,%.5g", ptf.x, YDIR(ptf.y));
-	    agset(n, "xlp", buf);
+	    agxbprint(&xb, "%.5g,%.5g", ptf.x, YDIR(ptf.y));
+	    agset(n, "xlp", agxbuse(&xb));
 	}
 	if (strcmp(ND_shape(n)->name, "record") == 0) {
 	    set_record_rects(n, ND_shape_info(n), &xb);
@@ -325,13 +325,13 @@ void attach_attrs_and_arrows(graph_t* g, int* sp, int* ep)
 		    if (i > 0)
 			agxbputc(&xb, ';');
 		    if (ED_spl(e)->list[i].sflag) {
-			s_arrows = 1;
+			s_arrows = true;
 			agxbprint(&xb, "s,%.5g,%.5g ",
 				ED_spl(e)->list[i].sp.x,
 				YDIR(ED_spl(e)->list[i].sp.y));
 		    }
 		    if (ED_spl(e)->list[i].eflag) {
-			e_arrows = 1;
+			e_arrows = true;
 			agxbprint(&xb, "e,%.5g,%.5g ",
 				ED_spl(e)->list[i].ep.x,
 				YDIR(ED_spl(e)->list[i].ep.y));
@@ -346,23 +346,23 @@ void attach_attrs_and_arrows(graph_t* g, int* sp, int* ep)
 		agset(e, "pos", agxbuse(&xb));
 		if (ED_label(e)) {
 		    ptf = ED_label(e)->pos;
-		    snprintf(buf, sizeof(buf), "%.5g,%.5g", ptf.x, YDIR(ptf.y));
-		    agset(e, "lp", buf);
+		    agxbprint(&xb, "%.5g,%.5g", ptf.x, YDIR(ptf.y));
+		    agset(e, "lp", agxbuse(&xb));
 		}
 		if (ED_xlabel(e) && ED_xlabel(e)->set) {
 		    ptf = ED_xlabel(e)->pos;
-		    snprintf(buf, sizeof(buf), "%.5g,%.5g", ptf.x, YDIR(ptf.y));
-		    agset(e, "xlp", buf);
+		    agxbprint(&xb, "%.5g,%.5g", ptf.x, YDIR(ptf.y));
+		    agset(e, "xlp", agxbuse(&xb));
 		}
 		if (ED_head_label(e)) {
 		    ptf = ED_head_label(e)->pos;
-		    snprintf(buf, sizeof(buf), "%.5g,%.5g", ptf.x, YDIR(ptf.y));
-		    agset(e, "head_lp", buf);
+		    agxbprint(&xb, "%.5g,%.5g", ptf.x, YDIR(ptf.y));
+		    agset(e, "head_lp", agxbuse(&xb));
 		}
 		if (ED_tail_label(e)) {
 		    ptf = ED_tail_label(e)->pos;
-		    snprintf(buf, sizeof(buf), "%.5g,%.5g", ptf.x, YDIR(ptf.y));
-		    agset(e, "tail_lp", buf);
+		    agxbprint(&xb, "%.5g,%.5g", ptf.x, YDIR(ptf.y));
+		    agset(e, "tail_lp", agxbuse(&xb));
 		}
 	    }
 	}
@@ -380,7 +380,7 @@ void attach_attrs_and_arrows(graph_t* g, int* sp, int* ep)
 
 void attach_attrs(graph_t * g)
 {
-    int e, s;
+    bool e, s;
     attach_attrs_and_arrows (g, &s, &e);
 }