+++ lib/common/utils.h
@@ -49,7 +49,7 @@ UTILS_API void free_queue(nodequeue *);
 UTILS_API void enqueue(nodequeue *, Agnode_t *);
 UTILS_API Agnode_t *dequeue(nodequeue *);
 UTILS_API pointf Bezier(pointf *, double, pointf *, pointf *);
-UTILS_API void attach_attrs_and_arrows(graph_t *, int *, int *);
+UTILS_API void attach_attrs_and_arrows(graph_t *, bool *, bool *);
 
 /** XML-escape a string
  *